pyproject.toml
[project]
name = "strands-agents-manim"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "loguru>=0.7.3",
    "manim>=0.19.1",
    "strands-agents>=0.1.0",
    "strands-agents-tools>=0.2.18",
]
